requirements.txt
anthropic
python-dotenv
fastmcp
openai
pydantic-ai
watchdog
